Don't lock executable location at load time
If a full path to the executable is frozen at load time, then setting exec-path later will confusingly fail to help the mode find the executable. This also prevents directory-local path mechanisms like direnv from working correctly.

A better default here is to use the unqualified executable name, and then the user can customise this to a full path if they need to.